+++ readenv/typing39.py
@@ -18,31 +18,12 @@
 # OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
 # THE SOFTWARE.
 
-import sys
-from typing import Tuple
+import re
 
-VersionType = Tuple[int, int, int, str, int]
+__all__ = [
+    "PatternType",
+    "MatchType",
+]
 
-PY39 = sys.version_info >= (3, 9)
-
-
-def get_version(version: VersionType) -> str:
-    """Returns a PEP 386-compliant version number from VERSION."""
-    assert len(version) == 5
-    assert version[3] in ("alpha", "beta", "rc", "final")
-
-    # Now build the two parts of the version number:
-    # main = X.Y[.Z]
-    # sub = .devN - for pre-alpha releases
-    #     | {a|b|c}N - for alpha, beta and rc releases
-
-    parts: int = 2 if version[2] == 0 else 3
-    main: str = ".".join(str(x) for x in version[:parts])
-
-    sub: str = ""
-
-    if version[3] != "final":
-        pre: str = {"alpha": "a", "beta": "b", "rc": "c"}[version[3]]
-        sub = f"{pre}{version[4]}"
-
-    return f"{main}{sub}"
+PatternType = re.Pattern[str]
+MatchType = re.Match[str]