+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## dev
+
+* fix typing support for python 3.8
+
 ## 0.1.1
 
 * can pass env absolute path